+++ MyApplication/Core/Repositories/ILearnedRepository.cs
@@ -5,8 +5,8 @@ namespace MyApplication.Core.Repositories
 {
     public interface ILearnedRepository
     {
-        bool CheckQuoteExistsInLearnedList(byte id, string userId);
-        Learned GetUserLearnedQuoteById(byte id, string userId);
+        bool CheckQuoteExistsInLearnedList(int id, string userId);
+        Learned GetUserLearnedQuoteById(int id, string userId);
         IEnumerable<Learned> GetUserLearnedQuotes(string userId);
         void Add(Learned learned);
         void Remove(Learned quote);
