+++ MyApplication/Core/Repositories/ILearningRepository.cs
@@ -5,9 +5,9 @@ namespace MyApplication.Core.Repositories
 {
     public interface ILearningRepository
     {
-        Learning GetUserLearningQuoteById(byte id, string userId);
+        Learning GetUserLearningQuoteById(int id, string userId);
         IEnumerable<Learning> GetUserLearningQuotes(string userId);
-        bool CheckQuoteExistsInLearnings(byte id, string userId);
+        bool CheckQuoteExistsInLearnings(int id, string userId);
         void Add(Learning learning);
         void Remove(Learning quote);
     }
